--- .env.sample
@@ -1,2 +1,0 @@
-NODE_ENV=development
-DATABASE_URL=postgres://aeviso:aeviso@postgres/aeviso
