+++ src/api/dev/usersRequest.http
@@ -10,7 +10,7 @@ content-type: application/json
 {
     "firstName": "Basile",
     "lastName": "Vernouillet",
-    "email": "bas@test.fr",
+    "email": "basi@test.fr",
     "role": "SUPERADMIN",
     "jobId": "e99f8391-1abc-4da6-98c4-6d1273c82c09",
     "companyId": "500c8dbd-ee35-44dc-92c9-a3598b9f9d63"
