+++ src/api/errors.js
@@ -3,8 +3,9 @@ const errors = {
     404: { message: "No User Found" },
     204: { message: "User Deleted" },
     P2025: { message: "No User Found" },
+    400: { message: "Missing Required Field" },
   },
-  400: { message: "Missing Required Field" },
+  400: { message: "Bad request" },
   P2002: { message: "Unique constraint failed" },
 };
 module.exports = errors;
