+++ src/api/index.js
@@ -1,10 +1,10 @@
 const express = require("express");
 
 const companies = require("./companies");
-const users = require("./users");
+const users = require("./users/routes");
 const records = require("./records");
 const professions = require("./professions");
-const projects = require("./projects")
+const projects = require("./projects");
 
 const router = express.Router();
 
@@ -18,6 +18,6 @@ router.use("/companies", companies);
 router.use("/users", users);
 router.use("/records", records);
 router.use("/professions", professions);
-router.use("/projects", projects)
+router.use("/projects", projects);
 
 module.exports = router;
