+++ src/api/users/controllers/post.js
@@ -1,0 +1,48 @@
+const errors = require("../../errors");
+const prisma = require("../../../../prismaClient");
+
+const post = async (req, res) => {
+  let company;
+  const {
+    firstName,
+    lastName,
+    email,
+    password,
+    role,
+    weeklyBasis,
+    companyId,
+    jobId,
+  } = req.body;
+  if (!firstName || !lastName || !email || !jobId) {
+    return res.status(400).json(errors.users[400]);
+  }
+  try {
+    if (companyId) {
+      company = {
+        connect: {
+          id: companyId,
+        },
+      };
+    }
+    const user = await prisma.user.create({
+      data: {
+        firstName,
+        lastName,
+        email,
+        password,
+        role,
+        weeklyBasis,
+        company,
+        job: {
+          connect: {
+            id: jobId,
+          },
+        },
+      },
+    });
+    res.status(201).json(user);
+  } catch (e) {
+    res.status(400).json(errors[e.code]);
+  }
+};
+module.exports = post;
